Fix build configuration for scikit-build 0.17
Switch from setting SKBUILD_CONFIGURE_OPTIONS to CMAKE_GENERATOR and set
the variable in the script invocation rather than in script_env.

Note the issue was not really with scikit-build 0.17 but with the
conda-forge packaging of it, which stopped listing `make` as a
dependency (which previous build were falling back to instead of using
`ninja`).

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>