Add nomkl to host dependencies (#55)
This will hopefully prevent the random build failures from mkl getting
installed and not handled correctly recently.

mkl has been getting installed into the host environment about 50% of
the time on Linux and leading to an overlinking error. I think that
means that the build is linking against mkl but not adding it to the run
exports, but I didn't test. Maybe an mkl build could work but for now
just block mkl since we had been doing fine building a non-mkl variant
of the package.